[IMP] DomEditable does not create 'editable' component.
The plugin add the feature of
contentEditable with event normalization but don't have any target by
default.
