+++ examples/odoo/odoo.ts
@@ -26,21 +26,16 @@ import {
     AlignJustifyButton,
 } from '../../packages/plugin-align/src/AlignButtons';
 import { DomLayout } from '../../packages/plugin-dom-layout/src/DomLayout';
-import { DomEditable } from '../../packages/plugin-dom-editable/src/DomEditable';
 import { OdooVideo } from '../../packages/plugin-odoo-video/src/OdooVideo';
 
 const target = document.getElementById('wrapper');
 target.innerHTML = template;
 
-const editor = new BasicEditor();
+const editor = new BasicEditor({ editable: target });
 editor.load(FontAwesome);
 editor.load(DevTools);
 editor.configure(DomLayout, {
     location: [target, 'replace'],
-});
-editor.configure(DomEditable, {
-    autoFocus: true,
-    source: target,
 });
 editor.configure(Toolbar, {
     layout: [
